improved my code added user controler

+++ NotesApplication/NotesApplication/Service/NoteService.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using NotesApplication.Models;
-using System.Drawing;
+using NotesApplication.Models.DTO;
+using NotesApplication.Models.Enums;
 
 namespace NotesApplication.Service
 {
@@ -13,61 +14,55 @@ namespace NotesApplication.Service
             this._context = context;
         }
 
-        public IQueryable<Note> GetAll()
+        public IQueryable<Note>GetAll()
         {
-            return _context.Notes.Include(n => n.User);
+            return _context.Notes;
         }
 
         public Note GetNote(int id)
         {
-            return _context.Notes.FirstOrDefault(n => n.Id == id);
+            return _context.Notes.Include(n => n.User).FirstOrDefault(n => n.Id == id);
         }
 
-        public IQueryable<Note> FilterBy (string filter)
+        public IQueryable<Note>FilterBy(string filter)
         {
-            return _context.Notes.Include(n => n.User).Where(n => n.Text.Contains(filter.ToLower()));
+            return _context.Notes.Where(n => n.Text.ToLower().Contains(filter.ToLower()));
         }
 
-        /*public IQueryable<Note> GetByName(string name)
+        public void Create(AddNoteDTO addNoteDTO)
         {
-            return _context.Notes.Include(n => n.User)
-                .Where(n => n.User.FirstName == name);
-        }*/
-
-        public void Create(Note note)
-        {
+            var note = new Note
+            {
+                Text = addNoteDTO.Text,
+                Color = Enum.Parse<Color>(addNoteDTO.Color, true),
+                UserId = 1
+            };
             _context.Notes.Add(note);
             _context.SaveChanges();
         }
 
-        public void Update(Note note)
+        public void Update(int id, AddNoteDTO addNoteDTO)
         {
+            var note = GetNote(id);
+            if (note != null)
+            {
+                note.Text = addNoteDTO.Text;
+                note.Color = Enum.Parse<Color>(addNoteDTO.Color, true);
+            }
+            throw new ArgumentException(message: "There is no note with that id");
+
             _context.Entry(note).State = EntityState.Modified;
-
-            try
-            {
-                _context.SaveChanges();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                throw;
-            }
-        }
-        public string ConvertEnumToString(Color color)
-        {
-            return color.ToString();
-        }
-
-        public Color ConvertStringToEnum(string color)
-        {
-            return (Color)Enum.Parse(typeof(Color), color);
-        }
-
-        public void DeleteNote(Note note)
-        {
-            _context.Notes.Remove(note);
             _context.SaveChanges();
         }
 
+        public void DeleteNote(int id)
+        {
+             var note = GetNote(id);
+            if (note == null)
+                throw new ArgumentException(message: "There is no note with that id");
+
+            _context.Notes.Remove(note);
+            _context.SaveChanges();
+        }
     }
 }
